Type Definition mentat_parser_utils::macros::ResultParser

type ResultParser<O, I> = Expected<FnParser<I, fn(_: I) -> ParseResult<O, I>>>;

A type definition for a function parser that either parses an O from an input stream of type I, or fails with an "expected" failure. See https://docs.rs/combine/2.2.1/combine/trait.Parser.html#method.expected for more illumination. Nothing about this is specific to the result type of the parser.