Struct mentat_parser_utils::value_and_span::OfExactly

pub struct OfExactly<P, N>(_, _);

OfExactly and of_exactly allow us to express nested parsers naturally.

For example, vector().of_exactly(many(list())) parses a vector-of-lists, like [(1 2) (:a :b) ("test") ()].

The "outer" parser P and the "nested" parser N must be compatible: P must produce an output edn::ValueAndSpan which can itself be turned into a stream of child elements; and N must accept the resulting input Stream. This compatibility allows us to lift errors from the nested parser to the outer parser, which is part of what has made parsing &'a [edn::Value] difficult.