Trait mentat_parser_utils::value_and_span::Item

pub trait Item<'a>: Clone + PartialEq + Sized {
    type Position: Clone + Ord + Display;
    fn start(&self) -> Self::Position;

    fn update_position(&self, _: &mut Self::Position);

    fn child_iter(&'a self) -> Iter<'a>;

    fn child_stream(&'a self) -> Stream<'a>;

    fn atom_iter(&'a self) -> Iter<'a>;

    fn atom_stream(&'a self) -> Stream<'a>;

    fn keyword_map_iter(&'a self) -> Iter<'a>;

    fn keyword_map_stream(&'a self) -> Stream<'a>;
}

Things specific to parsing with combine and our Stream that need a trait to live outside of the edn crate.

Associated Types

type Position: Clone + Ord + Display

Position could be specialized to SpanPosition.

Required Methods

fn start(&self) -> Self::Position

A slight generalization of combine::Positioner that allows to set the position based on the edn::ValueAndSpan being iterated.

fn update_position(&self, _: &mut Self::Position)

Important traits for Iter<'a>
impl<'a> Iterator for Iter<'a> type Item = &'a ValueAndSpan;

fn child_iter(&'a self) -> Iter<'a>

fn child_stream(&'a self) -> Stream<'a>

Important traits for Iter<'a>
impl<'a> Iterator for Iter<'a> type Item = &'a ValueAndSpan;

fn atom_iter(&'a self) -> Iter<'a>

fn atom_stream(&'a self) -> Stream<'a>

Important traits for Iter<'a>
impl<'a> Iterator for Iter<'a> type Item = &'a ValueAndSpan;

fn keyword_map_iter(&'a self) -> Iter<'a>

fn keyword_map_stream(&'a self) -> Stream<'a>